+++ src/config.ts
@@ -161,7 +161,7 @@ interface ConfigType {
  */
 class Config implements ConfigType {
     api = {
-        backgroundLoadSleepBetweenRequestsMS: 1_500, // How long to wait between API requests during backgrund load
+        backgroundLoadSleepBetweenRequestsMS: 1_000, // How long to wait between API requests during backgrund load
         backgroundLoadIntervalMinutes: 10,      // Time between background polling for additional user data after initial load
         daysBeforeFullCacheRefresh: 21,         // Days before the cache is considered stale and needs to be refreshed completely  // TODO: currently unused
         defaultRecordsPerPage: 40,              // Max per page is usually 40: https://docs.joinmastodon.org/methods/timelines/#request-2